Minor updates to Attributes.
Added more customization to the use of creating a PreferenceGroup from a class object so that the DefaultValue can be set from the current values if the PreferenceAttribute.DefaultValue is null. Also added additional methods to Preference and PreferenceGroup to streamline common activities of setting the values to null and to the DefaultValue.