Added Groups can be built with structs.
Previously, only a class's public members with types of classes or nullable structs (i.e., bool? and int?) could be built into Preferences. This is due to the nature of a Preference's Value and DefaultValue being inherently nullable, which is desirable since it allows for the default behavior of a null value means it has not be set. However, for struct members the behavior for setting a Value as null, it will use the DefaultValue, if not null, or the struct's default.
In addition, added the ability for updating a PreferenceFile to write after, which ensures the file is correct (e.g., comments and formatting).