Finished PreferenceAttribute implementation.
Completed the rest of implementing the PreferenceAttribute by allowing for the AllowedValues and SortAllowedValues to be set. The AllowUndefinedValues was refactored to only be set in the constructor parameter as a bool, but read as a bool?. This allows for knowing if the AllowUndefinedValues was specified or not in the constructor and if not the default behavior can be used for the specific Preference.
Also streamlined the naming of the ValidityProcessor as the ValueValidityProcessor was cumbersome.